ai_model/requirements.txt
unstructured
python-dotenv
pydantic
fastapi
uvicorn
requests
python-multipart
pdfplumber
geopy
google-cloud-documentai 
pandas
boto3
opencv-python
mediapipe 
numpy